requirements.txt
pyspark>=3.4.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
pytest>=6.2.0
python-dotenv>=0.19.0
kafka-python>=2.0.2
boto3>=1.26.0
jupyter>=1.0.0
scikit-learn>=1.0.0
imbalanced-learn>=0.8.0
pyarrow>=6.0.0
